Header: Increase the bounds of the Burger What's happening here? We want the sticky element to scale up and stick to the center of the burger. But while leaving we want to move the cursor away for some time before triggering the "onmouseleave" event.
To achieve this, we create a div inside of the burger menu that will
scale when we hover over it (the burger).
Now we have an element with a bigger bounds and it will not trigger
the onmouseleave event till we move the cursor outside the bounding box.

The `pointer-events:none` will stop triggering any events on the bounding box
and the bounding box will trigger ALL the pointer events because of `all`.

Now, let's move to making the cursor more "sticky".